- refactored with APIRouter
- fixed missing registration
- clened up in view of removing dynamic imports (to do)

- simplified module discovery

---
